Move everything except Popup to common
