+++ compose/material3/material3/src/commonMain/kotlin/androidx/compose/material3/ModalBottomSheet.kt
@@ -16,18 +16,48 @@
 
 package androidx.compose.material3
 
+import androidx.compose.animation.core.TweenSpec
+import androidx.compose.animation.core.animateFloatAsState
+import androidx.compose.foundation.Canvas
+import androidx.compose.foundation.gestures.Orientation
+import androidx.compose.foundation.gestures.detectTapGestures
+import androidx.compose.foundation.gestures.draggable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxWithConstraints
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.widthIn
 import androidx.compose.material3.SheetValue.Expanded
 import androidx.compose.material3.SheetValue.Hidden
+import androidx.compose.material3.SheetValue.PartiallyExpanded
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.graphics.isSpecified
+import androidx.compose.ui.input.nestedscroll.nestedScroll
+import androidx.compose.ui.input.pointer.pointerInput
+import androidx.compose.ui.semantics.clearAndSetSemantics
+import androidx.compose.ui.semantics.collapse
+import androidx.compose.ui.semantics.dismiss
+import androidx.compose.ui.semantics.expand
+import androidx.compose.ui.semantics.paneTitle
+import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.unit.Dp
-
+import androidx.compose.ui.unit.IntOffset
+import kotlin.math.max
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.launch
 
 /**
  * <a href="https://m3.material.io/components/bottom-sheets/overview" class="external" target="_blank">Material Design modal bottom sheet</a>.
@@ -62,7 +92,7 @@ import androidx.compose.ui.unit.Dp
  */
 @Composable
 @ExperimentalMaterial3Api
-expect fun ModalBottomSheet(
+fun ModalBottomSheet(
     onDismissRequest: () -> Unit,
     modifier: Modifier = Modifier,
     sheetState: SheetState = rememberModalBottomSheetState(),
@@ -74,7 +104,146 @@ expect fun ModalBottomSheet(
     dragHandle: @Composable (() -> Unit)? = { BottomSheetDefaults.DragHandle() },
     windowInsets: WindowInsets = BottomSheetDefaults.windowInsets,
     content: @Composable ColumnScope.() -> Unit,
-)
+) {
+    val scope = rememberCoroutineScope()
+    val animateToDismiss: () -> Unit = {
+        if (sheetState.swipeableState.confirmValueChange(Hidden)) {
+            scope.launch { sheetState.hide() }.invokeOnCompletion {
+                if (!sheetState.isVisible) {
+                    onDismissRequest()
+                }
+            }
+        }
+    }
+    val settleToDismiss: (velocity: Float) -> Unit = {
+        scope.launch { sheetState.settle(it) }.invokeOnCompletion {
+            if (!sheetState.isVisible) onDismissRequest()
+        }
+    }
+
+    // Callback that is invoked when the anchors have changed.
+    val anchorChangeHandler = remember(sheetState, scope) {
+        ModalBottomSheetAnchorChangeHandler(
+            state = sheetState,
+            animateTo = { target, velocity ->
+                scope.launch { sheetState.animateTo(target, velocity = velocity) }
+            },
+            snapTo = { target ->
+                val didSnapImmediately = sheetState.trySnapTo(target)
+                if (!didSnapImmediately) {
+                    scope.launch { sheetState.snapTo(target) }
+                }
+            }
+        )
+    }
+
+    ModalBottomSheetPopup(
+        onDismissRequest = {
+            if (sheetState.currentValue == Expanded && sheetState.hasPartiallyExpandedState) {
+                scope.launch { sheetState.partialExpand() }
+            } else { // Is expanded without collapsed state or is collapsed.
+                scope.launch { sheetState.hide() }.invokeOnCompletion { onDismissRequest() }
+            }
+        },
+        windowInsets = windowInsets,
+    ) {
+        BoxWithConstraints(Modifier.fillMaxSize()) {
+            val fullHeight = constraints.maxHeight
+            Scrim(
+                color = scrimColor,
+                onDismissRequest = animateToDismiss,
+                visible = sheetState.targetValue != Hidden
+            )
+            val bottomSheetPaneTitle = getString(string = Strings.BottomSheetPaneTitle)
+            Surface(
+                modifier = modifier
+                    .widthIn(max = BottomSheetMaxWidth)
+                    .fillMaxWidth()
+                    .align(Alignment.TopCenter)
+                    .semantics { paneTitle = bottomSheetPaneTitle }
+                    .offset {
+                        IntOffset(
+                            0,
+                            sheetState
+                                .requireOffset()
+                                .toInt()
+                        )
+                    }
+                    .nestedScroll(
+                        remember(sheetState) {
+                            ConsumeSwipeWithinBottomSheetBoundsNestedScrollConnection(
+                                sheetState = sheetState,
+                                orientation = Orientation.Vertical,
+                                onFling = settleToDismiss
+                            )
+                        }
+                    )
+                    .modalBottomSheetSwipeable(
+                        sheetState = sheetState,
+                        anchorChangeHandler = anchorChangeHandler,
+                        screenHeight = fullHeight.toFloat(),
+                        onDragStopped = {
+                            settleToDismiss(it)
+                        },
+                    ),
+                shape = shape,
+                color = containerColor,
+                contentColor = contentColor,
+                tonalElevation = tonalElevation,
+            ) {
+                Column(Modifier.fillMaxWidth()) {
+                    if (dragHandle != null) {
+                        val collapseActionLabel =
+                            getString(Strings.BottomSheetPartialExpandDescription)
+                        val dismissActionLabel = getString(Strings.BottomSheetDismissDescription)
+                        val expandActionLabel = getString(Strings.BottomSheetExpandDescription)
+                        Box(
+                            Modifier
+                                .align(Alignment.CenterHorizontally)
+                                .semantics(mergeDescendants = true) {
+                                    // Provides semantics to interact with the bottomsheet based on its
+                                    // current value.
+                                    with(sheetState) {
+                                        dismiss(dismissActionLabel) {
+                                            animateToDismiss()
+                                            true
+                                        }
+                                        if (currentValue == PartiallyExpanded) {
+                                            expand(expandActionLabel) {
+                                                if (swipeableState.confirmValueChange(Expanded)) {
+                                                    scope.launch { sheetState.expand() }
+                                                }
+                                                true
+                                            }
+                                        } else if (hasPartiallyExpandedState) {
+                                            collapse(collapseActionLabel) {
+                                                if (
+                                                    swipeableState.confirmValueChange(
+                                                        PartiallyExpanded
+                                                    )
+                                                ) {
+                                                    scope.launch { partialExpand() }
+                                                }
+                                                true
+                                            }
+                                        }
+                                    }
+                                }
+                        ) {
+                            dragHandle()
+                        }
+                    }
+                    content()
+                }
+            }
+        }
+    }
+    if (sheetState.hasExpandedState) {
+        LaunchedEffect(sheetState) {
+            sheetState.show()
+        }
+    }
+}
 
 /**
  * Create and [remember] a [SheetState] for [ModalBottomSheet].
@@ -90,3 +259,104 @@ fun rememberModalBottomSheetState(
     skipPartiallyExpanded: Boolean = false,
     confirmValueChange: (SheetValue) -> Boolean = { true },
 ) = rememberSheetState(skipPartiallyExpanded, confirmValueChange, Hidden)
+
+@Composable
+private fun Scrim(
+    color: Color,
+    onDismissRequest: () -> Unit,
+    visible: Boolean
+) {
+    if (color.isSpecified) {
+        val alpha by animateFloatAsState(
+            targetValue = if (visible) 1f else 0f,
+            animationSpec = TweenSpec()
+        )
+        val dismissSheet = if (visible) {
+            Modifier
+                .pointerInput(onDismissRequest) {
+                    detectTapGestures {
+                        onDismissRequest()
+                    }
+                }
+                .clearAndSetSemantics {}
+        } else {
+            Modifier
+        }
+        Canvas(
+            Modifier
+                .fillMaxSize()
+                .then(dismissSheet)
+        ) {
+            drawRect(color = color, alpha = alpha)
+        }
+    }
+}
+
+@ExperimentalMaterial3Api
+private fun Modifier.modalBottomSheetSwipeable(
+    sheetState: SheetState,
+    anchorChangeHandler: AnchorChangeHandler<SheetValue>,
+    screenHeight: Float,
+    onDragStopped: CoroutineScope.(velocity: Float) -> Unit,
+) = draggable(
+        state = sheetState.swipeableState.swipeDraggableState,
+        orientation = Orientation.Vertical,
+        enabled = sheetState.isVisible,
+        startDragImmediately = sheetState.swipeableState.isAnimationRunning,
+        onDragStopped = onDragStopped
+    )
+    .swipeAnchors(
+        state = sheetState.swipeableState,
+        anchorChangeHandler = anchorChangeHandler,
+        possibleValues = setOf(Hidden, PartiallyExpanded, Expanded),
+    ) { value, sheetSize ->
+        when (value) {
+            Hidden -> screenHeight
+            PartiallyExpanded -> when {
+                sheetSize.height < screenHeight / 2 -> null
+                sheetState.skipPartiallyExpanded -> null
+                else -> screenHeight / 2f
+            }
+            Expanded -> if (sheetSize.height != 0) {
+                max(0f, screenHeight - sheetSize.height)
+            } else null
+        }
+    }
+
+@ExperimentalMaterial3Api
+private fun ModalBottomSheetAnchorChangeHandler(
+    state: SheetState,
+    animateTo: (target: SheetValue, velocity: Float) -> Unit,
+    snapTo: (target: SheetValue) -> Unit,
+) = AnchorChangeHandler<SheetValue> { previousTarget, previousAnchors, newAnchors ->
+    val previousTargetOffset = previousAnchors[previousTarget]
+    val newTarget = when (previousTarget) {
+        Hidden -> Hidden
+        PartiallyExpanded, Expanded -> {
+            val hasPartiallyExpandedState = newAnchors.containsKey(PartiallyExpanded)
+            val newTarget = if (hasPartiallyExpandedState) PartiallyExpanded
+            else if (newAnchors.containsKey(Expanded)) Expanded else Hidden
+            newTarget
+        }
+    }
+    val newTargetOffset = newAnchors.getValue(newTarget)
+    if (newTargetOffset != previousTargetOffset) {
+        if (state.swipeableState.isAnimationRunning || previousAnchors.isEmpty()) {
+            // Re-target the animation to the new offset if it changed
+            animateTo(newTarget, state.swipeableState.lastVelocity)
+        } else {
+            // Snap to the new offset value of the target if no animation was running
+            snapTo(newTarget)
+        }
+    }
+}
+
+/**
+ * Popup specific for modal bottom sheet.
+ */
+@Composable
+internal expect fun ModalBottomSheetPopup(
+    onDismissRequest: () -> Unit,
+    windowInsets: WindowInsets,
+    content: @Composable () -> Unit,
+)
